fix: Ptr{Cvoid}(UInt32) on Windows — use UInt(id) for pointer conversion
On Windows, thread_id_t is UInt32 but Ptr{Cvoid}() only accepts
UInt64/Int64/Ptr. Convert via UInt() to get pointer-sized integer.

+++ src/foreign_threads.jl
@@ -88,7 +88,7 @@ function ForeignThread(name::String, thread_fn::Ref{Ptr{Cvoid}};
             ret = ccall(
                 (:CreateThread, "kernel32"), Ptr{Cvoid},
                 (Ptr{Cvoid}, Csize_t, Ptr{Cvoid}, Ptr{Cvoid}, UInt32, Ptr{UInt32}),
-                C_NULL, Csize_t(0), thread_fn[], Ptr{Cvoid}(id), UInt32(0), C_NULL,
+                C_NULL, Csize_t(0), thread_fn[], Ptr{Cvoid}(UInt(id)), UInt32(0), C_NULL,
             )
             ret == C_NULL && throw(ArgumentError("error creating OS thread for ForeignThread"))
             # Close handle — thread keeps running, OS auto-cleans on exit
